refactor(github): split out useGithubFileForm
